ept_array_lastpostag changed
In the "run model" code block, the last line of the save heads for loop, I changed ept_array_lastpostdev to ept_array_lastpostag...not sure if this should have stayed "...lastpostdev" but there were no lines creating the object.

Will need to run the code and see what it does.