NEW CODE: Breanna
Delayed upload for almost a week.

As part of an effort to better understand the coding process of this thesis, I am rewriting portions of the code: primarily by simplifying the lines with for loops.

This latest commit focuses on the nonbehavioralModels_fn. I will replace IF statements for the IN_BASIS variable with the dictionary variables.