Completed: truth data likelihood record
This function records the truth values from the truth heads/flows files and creates an object called truth_value. 

It also respond to a flag (useTrueData) that allows this step to be bypassed.

The next portion of coding focuses on creating a function that assigns model likelihood based on the mismatch to true values. A second function seated within will need to be developed that assigns a huge rmse value to nonbehavioral models so that they can be easily dropped later on.