In progress: generate head_related contour plots (cont'd)
Replaced variable names. Still getting an error that says 'arange: cannot compute length'

It has something to do with the function Ty made which generates plots. This is the last plot in the cell block and it gets stuck. Must be due to an np.arange() method. 

Try going back through Ty's function and look for an np.arange() method.